Make Stats barchart Y-Axis start at 0
This commit forces the Y-axis of the barchart on the user stats page
to start at 0. Previously, the Y-axis would start at a
dynamically-calculated value based on the data in the chart.
This meant that the bars did not work as a visual
representation of the data, because a bar half the size of another
bar did not depict a value half the size of another value.

So eg:
if bars depicted hitcount, and the Y-axis started at 50, a fic with
100 hits would have a bar double the size of a fic with 75 hits.

This patch forces the Y-axis to start at 0, by setting the
Google Chart API `vAxis: {Minvalue}` parameter to 0.

(WIP; filed report but no ticket num yet, but tested the code on a
local dev instance and it works there. Will delete this, shorten
the commit message and add ticket num to the title if/when
there's a ticket. :) )